% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candi-package.R
\docType{package}
\name{candi}
\alias{candi}
\alias{candi-package}
\alias{candi-package}
\title{candii}
\description{
Demonstrate shiny candi radiograph dashboard implementations.
}
\author{
\strong{Maintainer}: Marcus Badgeley \email{mabadgeley@gmail.com}

}
