% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{candiOpt_}
\alias{candiOpt_}
\alias{candiOpt}
\title{Convenience fxn to fetch an option set by the hips package.}
\usage{
candiOpt_(opt_substr, ...)

candiOpt(opt_substr, ...)
}
\arguments{
\item{opt_substr}{chr(1) specific hips option}

\item{...}{optional default if variable isn't set \code{\link[base]{options}}}
}
\value{
set option or ...
}
\description{
candiOpt uses NSE to save typing, use candiOpt_ for SE
}
\details{
Package option schema:
[pkg name] DOT [snake_case_option_name]
eg: hips.option_one
}
\examples{
candiOpt_("annotation_types")
candiOpt()
candiOpt(annotation_types)
}
