% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_images.R
\name{case}
\alias{case}
\alias{load_case}
\alias{display_case}
\title{Load radiographs from dataset-case into an EBImage stack}
\usage{
load_case(dataset, case_id, img_dir = kIMG_DIR)

display_case(dataset, case_id, img_dir = kIMG_DIR)
}
\arguments{
\item{dataset}{chr(1)}

\item{case_id}{chr(1)}

\item{img_dir}{chr(1)}
}
\value{
\code{\link[EBImage]{Image}} with 4 axes
}
\description{
All images named <dataset>_<case_id>_* in img_dir are loaded.
}
\examples{
load_case("iu", "1") \%>\% EBImage::display(.)
display_case("iu", "3")
}
