% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{digest_user_to_id}
\alias{digest_user_to_id}
\title{Convert a username into a coded integer ID using sha256 digests trimmed to 4 bytes}
\usage{
digest_user_to_id(username)
}
\arguments{
\item{username}{chr(1) user name with correct case}
}
\value{
a 4 or 5 digit integer
}
\description{
Convert a username into a coded integer ID using sha256 digests trimmed to 4 bytes
}
\examples{
digest_user_to_id("Marcus")
digest_user_to_id("Nick")
map_int(letters, digest_user_to_id)
}
