% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_emr.R
\name{emrModule}
\alias{emrModule}
\alias{hpiOutput}
\alias{hpiModule}
\alias{histImpOutput}
\alias{histImpModule}
\title{Scalar data display modules for shiny apps}
\usage{
hpiOutput(id)

hpiModule(input, output, session, idIn)

histImpOutput(id)

histImpModule(input, output, session, idIn)
}
\arguments{
\item{id}{chr(1) namespace for the module, decided by the caller at the time the module is used}

\item{input, output, session}{standard args to shiny server routine}

\item{idIn}{reactive that returns a chr(1) case or image identifier to fetch data for}
}
\value{
\code{\link[shiny]{tagList}} of emr ui components
}
\description{
Scalar data display modules for shiny apps
}
\examples{
if (interactive()) {
library(shiny)

shinyApp(
    ui = fluidPage(
        selectInput("idIn", "Case ID:", choices = candi::radiographs$img_id),
        hpiOutput("hpi")
    ),
    server = function(input, output, session) {
        callModule(hpiModule, "hpi", idIn = reactive(input$idIn))
    }
)

shinyApp(
    ui = fluidPage(
        selectInput("idIn", "Case ID:", choices = radiographs$img_id),
        histImpOutput("histImp")
    ),
    server = function(input, output, session) {
        callModule(histImpModule, "histImp", idIn = reactive(input$idIn))
    }
)
}
}
