% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bien_occ_records.R
\name{get_bien_occ_records}
\alias{get_bien_occ_records}
\title{get bien records}
\usage{
get_bien_occ_records(sp)
}
\arguments{
\item{sp}{A species name. First letter of genus names are capitalized and specific epithet is lowercase with a space separating the two.}
}
\value{
A dataframe of occurrence data for species sp
}
\description{
A function to download occurrence data from the BIEN database for a given list of species
}
\examples{
get_bien_occ_records(c("Notholaena standleyi","Pellaea truncata"))
}
