% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_both_occ_records.R
\name{get_both_occ_records}
\alias{get_both_occ_records}
\title{get both gbif and bien records and combine}
\usage{
get_both_occ_records(sp, limit = 500)
}
\arguments{
\item{sp}{A species name. First letter of genus names are capitalized and specific epithet is lowercase with a space separating the two.}

\item{limit}{A threshold of maximum number of occurrences to download. Default set to 500}
}
\value{
A dataframe of occurrence data for species sp
}
\description{
A function to download occurrence data for a species from both the BIEN and GBIF database and combine the records
}
\examples{
get_both_occ_records("Notholaena standleyi", limit = 100)
}
