% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_occ_records.R
\name{get_gbif_occ_records}
\alias{get_gbif_occ_records}
\title{get gbif records}
\usage{
get_gbif_occ_records(sp, limit = 500)
}
\arguments{
\item{sp}{A species name. First letter of genus names are capitalized and specific epithet is lowercase with a space separating the two.}

\item{limit}{A threshold of maximum number of occurrences to download. Default set to 500}
}
\value{
A dataframe of occurrence data for species sp
}
\description{
A function to download occurrence data from the GBIF database for a given list of species
}
\examples{
get_gbif_occ_records(c("Notholaena standleyi","Pellaea truncata"), limit = 100)
}
