% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occ_records.R
\name{get_occ_records}
\alias{get_occ_records}
\title{Get occurrence records}
\usage{
get_occ_records(species, database, limit = 500)
}
\arguments{
\item{species}{A list of species separated by spaces. First letter of genus names are capitalized and specific epithet is lowercase with a space separating the two.}

\item{database}{User specifies if they want occurrence data downloaded from BIEN, GBIF, or both. Value must be "bien", "gbif", or "both"}

\item{limit}{A threshold of maximum number of occurrences to download. Default set to 500}
}
\value{
A list of dataframes of occurrence data for each species in species
}
\description{
A function that downloads occurrence records for each species in a list of species from either BIEN, GBIF, or both databases
}
\examples{
get_occ_records(c("Notholaena standleyi","Pellaea truncata", "Astrolepis integerrima"), "both", limit = 500)
}
