% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_data_functions.R
\name{get_world_clim}
\alias{get_world_clim}
\title{get world clim}
\usage{
get_world_clim()
}
\value{
A raster stack of the 19 bioclimatic world clim variables
}
\description{
A function to download the 19 bioclimatic variables from the world clim database
}
