% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio.R
\name{getLastCaseChr}
\alias{getLastCaseChr}
\title{Utility fxn to search all user input data to identify the most recent case reviewed by a user}
\usage{
getLastCaseChr(user_name, usr_input_dir)
}
\arguments{
\item{user_name}{chr(1) must match a col user_name in the user input data frames}

\item{usr_input_dir}{chr(1) directory containing user input records}
}
\value{
chr(1) img_id of last image reviewed
}
\description{
This function is used by the CANDI_CAD trial state manager to reserve an image
for second reader mode after initially having the user review an image without assistance.
}
\examples{
getLastCaseChr(user_name="Marcus", usr_input_dir="/www/app_data_candi/CANDI_CAD/usr_inpt")
}
