% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio.R
\name{getNewCaseDf}
\alias{getNewCaseDf}
\title{Randomly select CAD state variables for a new test image: test_img_id and reader_mode}
\usage{
getNewCaseDf(test_img_ids)
}
\description{
Randomly select CAD state variables for a new test image: test_img_id and reader_mode
}
