% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio.R
\name{handle_annotation_download}
\alias{handle_annotation_download}
\title{Create a download handler for one of the annotation types in \code{kANNOTATION_TYPES}}
\usage{
handle_annotation_download(ann_type, f_load)
}
\arguments{
\item{ann_type}{chr(1), one of the globally assigned}

\item{f_load}{function to load annotations such as \code{\link{load_gs_annotation}}}
}
\value{
assign return to a slot on shiny's \code{output} and in the UI use \code{\link[shiny]{downloadButton}} or \code{\link[shiny]{downloadLink}}
}
\description{
Create a download handler for one of the annotation types in \code{kANNOTATION_TYPES}
}
