% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_imageOutput.R
\name{image_invert}
\alias{image_invert}
\title{Invert the color of an image.}
\usage{
image_invert(Img)
}
\arguments{
\item{Img}{an \code{\link[EBImage]{Image}}}
}
\value{
an \code{\link[EBImage]{Image}}
}
\description{
Invert the color of an image.
}
