% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ids.R
\name{img_id}
\alias{img_id}
\alias{sep_id}
\alias{unite_id}
\alias{with_sep}
\title{Image Identifier Helper Transformations}
\usage{
sep_id(df, img_id_cmpd = kIMG_ID_S_CMPD, img_id_vars = kIMG_ID_X_VARS,
  sep = "_", remove = TRUE, convert = TRUE, ...)

unite_id(df, img_id_cmpd = kIMG_ID_S_CMPD, img_id_vars = kIMG_ID_X_VARS,
  sep = "_", remove = TRUE)

with_sep(df, f, ...)
}
\arguments{
\item{df}{data.frame with column 'img_id' or columns `img_vars`}

\item{img_id_cmpd}{s_chr compound key identifier representation}

\item{img_id_vars}{x_chr names of individual tidy id variables}

\item{...}{args passed to \code{\link[tidyr]{separate}} or \code{\link[tidyr]{unite}}}

\item{f}{function to call on df after separating ids}
}
\value{
data.frame with transformed id column representation
}
\description{
Image Ids can be represented in 2 ways:
\enumerate{
    \item chr(1) unique compound key
    \item chr(n) vector of individual id component variables
}
}
\details{
The default names of id vars are configured with the global `kID_CHRS`,
which is defined in this img id helpers script

Mutating Fxn to separate `img_id` column of a data.frame into separate component identifiers.

Assumes the project schema of img_id=<dataset>_<case>_<image>
Attempts to attach original attributes onto the product (except names and spec)

dataset, case, image cols.

Unite id components into a single unique image identifier.

Convenience wrapper to directly operate on a joint id as if it were split
}
\examples{
data(lut_img_id2original)
sep_id(lut_img_id2original)
sep_id(lut_img_id2original, remove=FALSE)
try(sep_id(bind_rows(lut_img_id2original, lut_img_id2original)))  # Expect warning
}
