% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imgIds2Cases}
\alias{imgIds2Cases}
\title{Convert a vector of image IDs to corresponding unique case IDs.}
\usage{
imgIds2Cases(img_ids)
}
\description{
Convert a vector of image IDs to corresponding unique case IDs.
}
\examples{
imgs_avail = candiOpt(large_img_dir) \%>\% list.files() \%>\% fp_stem()
imgIds2Cases(imgs_avail)
}
