% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_impressionInput.R
\name{impressionInput}
\alias{impressionInput}
\title{Impression User Interface Module}
\usage{
impressionInput(id, dx_chr, include_demographics = FALSE,
  include_technical = FALSE)
}
\arguments{
\item{id}{chr(1) namespace for the module, decided by the caller at the time the module is used}

\item{dx_chr}{chr(n) diagnoses to include in the classification input}
}
\value{
\code{\link[shiny]{tagList}} of impression ui components
}
\description{
Create a collection of ui inputs for a user to indicate binary classification of
each dx in \code{dx_chr} along with a free text note.
}
\examples{
if (interactive()) {
library(shiny)

include_demographics <- TRUE
include_technical <- TRUE

shinyApp(
    ui = fluidPage(fluidRow(
        column(6,
            impressionInput("impression", dx_chr = candiOpt(dxs_chr),
                            include_demographics = include_demographics,
                            include_technical = include_technical)
        ),
        column(6,
            shiny::tags$strong("User Impression:"),
            tableOutput("usrImpressionTable")
        )
    )),
    server = function(input, output) {
        usrImpressionDf <- callModule(impression, "impression",
                            include_demographics = include_demographics,
                            include_technical = include_technical)
        )
        output$usrImpressionTable <- renderTable(usrImpressionDf())
    }
)}
}
\seealso{
\code{\link{impression}}

Other shiny_module: \code{\link{impression}},
  \code{\link{radiographOutput}},
  \code{\link{similarImgUi}}, \code{\link{similarImg}}
}
