% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impressionUi.R
\name{impressionUi}
\alias{impressionUi}
\title{Shiny Module User Impression Input Interface}
\usage{
impressionUi(id, dx_chr)
}
\arguments{
\item{id}{chr(1) namespace for the module, decided by the caller at the time the module is used}

\item{dx_chr}{chr(n) diagnosis to include in the classification input}
}
\value{
\code{\link[shiny]{tagList}} of impression ui components
}
\description{
Shiny Module User Impression Input Interface
}
\seealso{
\code{\link{impression}}

Other shiny_module: \code{\link{impression}}
}
