% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ids.R
\name{isImgId}
\alias{isImgId}
\alias{isCaseId}
\title{Validation functions for flexibly handling different id scoping}
\usage{
isImgId(x_chr)

isCaseId(x_chr)
}
\arguments{
\item{x_chr}{character vector to test for id format}
}
\description{
Validation functions for flexibly handling different id scoping
}
\examples{
isImgId("iu_1_1")
isImgId("iu")
isImgId(1)
isImgId(c("iu_1_1", "iu_4_4"))
isCaseId("iu_1_1")
isCaseId("iu")
isCaseId(1)
isCaseId("1")
isCaseId(c("1", "5", "5000"))
}
