% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio_image.R
\name{load_case}
\alias{load_case}
\alias{display_case}
\title{Load radiographs from dataset-case into an EBImage stack}
\usage{
load_case(case_id, dataset = "iu", img_dir = candiOpt(large_img_dir))

display_case(case_id, dataset = "iu", img_dir = candiOpt(large_img_dir))
}
\arguments{
\item{case_id}{chr(1) or int(1)}

\item{dataset}{chr(1) default "iu"}

\item{img_dir}{chr(1) default candiOpt(large_img_dir)}
}
\value{
\code{\link[EBImage]{Image}} with 4 axes
}
\description{
All images named <dataset>_<case_id>_* in img_dir are loaded.
}
\examples{
load_case("1") \%>\% EBImage::display(.)
load_case(1) \%>\% EBImage::display(.)
display_case("3")
}
