% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio_usrDownloads.R
\name{load_csv_annotation}
\alias{load_csv_annotation}
\title{Load csv files for a given annotation type}
\usage{
load_csv_annotation(dir, ann_type)
}
\description{
Load csv files for a given annotation type
}
\examples{
load_csv_annotation(kANN_DIR, "classification")
}
