% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio.R
\name{load_usr_input}
\alias{load_usr_input}
\title{Load input app data from a user}
\usage{
load_usr_input(user, usr_inpt_dir)
}
\arguments{
\item{user}{chr(1) user to filter user_name column by}
}
\value{
data.frame
}
\description{
By default, data is loaded form a directory indicated by a
global variable defined in the file that defines \code{load_usr_input}.
}
\examples{
\dontrun{
  load_usr_input("Marcus", "/www/app_data_cxrTargetDiff/usr_inpt/")
}
}
