% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio_usrInput.R
\name{log_usr_event}
\alias{log_usr_event}
\title{Save a timestamped record when a user triggers an event.}
\usage{
log_usr_event(usr_chr, event_chr, dir, img_id = NA_character_)
}
\arguments{
\item{usr_chr}{chr(1) username}

\item{event_chr}{chr(1) event name}

\item{dir}{chr(1)}

\item{img_id}{optional chr(1)}
}
\value{
target file path, but called for saving side effect
}
\description{
Save a timestamped record when a user triggers an event.
}
\examples{
save_fp <- log_usr_event("marcus", "submitBtn", dir = candiOpt(app_data_dir))
read_csv(save_fp); file.remove(save_fp)            # and remove
}
