% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corr_matrix.R
\name{make_corr_matrix}
\alias{make_corr_matrix}
\title{make corr matrix}
\usage{
make_corr_matrix(occurrences, environment_data)
}
\arguments{
\item{environment_data}{A raster stack of environmental raster data variables}

\item{ocurrences}{A list of data frames with the occurrence data for each species}
}
\value{
a pdf of the correlation matrix is saved to the local disk
}
\description{
A functiont that creates a correlation matrix between the environmental data variables of correlation within the area of the maximum extents of occurrence points
}
