% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_niches.R
\name{model_niches}
\alias{model_niches}
\title{model niches}
\usage{
model_niches(
  occ_data,
  enviro_data,
  reps = 3,
  reptype = "bootstrap",
  output_dir = ""
)
}
\arguments{
\item{occ_data}{A list of dataframes where each element of the list os a dataframe of occurrence data for an individual species}

\item{enviro_data}{A rasterstack of environmental rasters. Can be continuous or categorical data}

\item{reps}{The number of replicates done when modeling. Default is 3}

\item{reptype}{specifies the Maxent replicatetype to use.Can be "bootstrap", "subsample", or "crossvalidate". Default set to "bootstrap"}

\item{output_dir}{The output directory of the returned object}
}
\value{
A list of maxent objects for each species and the associated raster of each model
}
\description{
A function to create maxent species distribution models for large numbers of taxa.
}
