% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mtx2idf}
\alias{mtx2idf}
\title{Convert a matrix with rownames into a column indexed data.frame}
\usage{
mtx2idf(mtx, id_col = "id")
}
\arguments{
\item{mtx}{matrix with rownames}
}
\value{
data.frame with column "id" containing mtx's rownames
}
\description{
Convert a matrix with rownames into a column indexed data.frame
}
\examples{
as.matrix(mtcars) \%>\% mtx2idf()
}
