% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_emr.R
\name{patientMedicalRecordOutput}
\alias{patientMedicalRecordOutput}
\alias{patientMedicalRecord}
\title{EMR display module for shiny apps}
\usage{
patientMedicalRecordOutput(id, section_label = "Patient Medical Record")

patientMedicalRecord(input, output, session, idIn)
}
\arguments{
\item{id}{chr(1) namespace for the module, decided by the caller at the time the module is used}

\item{section_label}{chr(1) text to print at the top of the ui section}
}
\value{
\code{\link[shiny]{tagList}} of emr ui components
}
\description{
EMR display module for shiny apps
}
\examples{
if (interactive()) {
shinyApp(
    ui = fluidPage(
        selectInput("idIn", "Case ID", choices = cases$case),
        patientMedicalRecordOutput("emr")
    ),
    server = function(input, output, session) {
        callModule(patientMedicalRecord, "emr", idIn = reactive(input$idIn))
    }
)}
}
