% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_images.R
\name{radiograph}
\alias{radiograph}
\alias{load_radiograph}
\alias{display_radiograph}
\title{Load and Display radiographs via EBImage Images}
\usage{
load_radiograph(img_id, img_dir = kIMG_DIR)

display_radiograph(img_id, img_dir = kIMG_DIR)
}
\arguments{
\item{img_id}{chr(1)}

\item{img_dir}{chr(1)}
}
\value{
\code{\link[EBImage]{Image}} with 3 axes
}
\description{
Wrapper for \code{\link[EBImage]{readImage}}
}
\examples{
load_radiograph("iu_1_1")
load_radiograph("iu_1_1") \%>\% EBImage::display(.)
display_radiograph("iu_1_1")
}
