% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_imageOutput.R
\name{radiographOutput}
\alias{radiographOutput}
\alias{radiograph}
\alias{radiograph}
\title{Radiograph display modules for shiny apps}
\usage{
radiographOutput(id, height = "1000px", width = "80\%")

radiograph(input, output, session, imgIdIn)
}
\arguments{
\item{id}{chr(1) namespace for the module, decided by the caller at the time the module is used}

\item{input, output, session}{shiny server io}

\item{imgIdIn}{reactive evaluating to a chr(1) with the file stem of the desired image}
}
\value{
shiny module that outputs the requested image
}
\description{
Wraps \code{\link[EBImage]{displayOutput}} in UI fxn
and \code{\link{display_radiograph}} in server fxn
}
\examples{
if (interactive()) {
library(shiny)

imgs_avail = candiOpt(large_img_dir) \%>\% list.files() \%>\% fp_stem()

shinyApp(
ui = fluidPage(
    selectInput("imgIdIn", "Image Id:", choices = imgs_avail),
    verbatimTextOutput("selectedImgTxt"),
    radiographOutput("show_rad")
),
server = function(input, output, session) {
    output$selectedImgTxt <- renderPrint(input$imgIdIn)
    callModule(radiograph, "show_rad", imgIdIn = reactive(input$imgIdIn))
})
}
}
\seealso{
Other shiny_module: \code{\link{impressionInput}},
  \code{\link{impression}}, \code{\link{similarImgUi}},
  \code{\link{similarImg}}
}
