% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candi-package.R
\docType{data}
\name{radiographs}
\alias{radiographs}
\title{radiograph-major data.frame with technical specs and CNN PCs and pYs}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7309 rows and 22 columns.}
\usage{
radiographs
}
\description{
radiograph-major data.frame with technical specs and CNN PCs and pYs
}
\keyword{datasets}
