% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{randomize_user_queue}
\alias{randomize_user_queue}
\title{Deterministically randomize a queue order for a user}
\usage{
randomize_user_queue(username, input_ids)
}
\arguments{
\item{username}{chr(1)}

\item{input_ids}{int(1)}
}
\value{
a vector the same length as input_ids, with a deterministically shuffled order
}
\description{
Deterministically randomize a queue order for a user
}
\examples{
randomize_user_queue("marcus", letters)
}
