% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_corr_variables.R
\name{remove_corr_variables}
\alias{remove_corr_variables}
\title{remove variables}
\usage{
remove_corr_variables(raster_stack, variables_to_be_removed)
}
\arguments{
\item{raster_stack}{A raster stack generated by the get_world_clim function}

\item{variables_to_be_removed}{A list of variables selected by the user}
}
\description{
A function that removes specified variables from the raster stack of climate variables
}
\examples{
remove_corr_variables(raster_stack, c("bio1", "bio4", "bio10"))
}
