% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dup_locs.R
\name{remove_dup_locs}
\alias{remove_dup_locs}
\title{remove duplicate localities}
\usage{
remove_dup_locs(df_list)
}
\arguments{
\item{df_list}{A list of data frames with latitude and longitude data for each species. Object generated from CANDI occurrence data aquisition functions}
}
\value{
The same list of dataframes with the duplicate localities removed
}
\description{
A function to remove duplicates latitude and longitudes
}
