% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_lessthan.R
\name{remove_lessthan}
\alias{remove_lessthan}
\title{remove less than}
\usage{
remove_lessthan(df_list, n = 5)
}
\arguments{
\item{df_list}{A list of data frames with latitude and longitude data for each species.}

\item{n}{the threshold of minimum number of occurrence records a dataframe must have to not be removed from the list}
}
\value{
The same list of data frames with any elements in the list removed that had less than n records
}
\description{
remove data frames in the list with fewer than n occurrence records
}
