% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_null_items.R
\name{remove_null_items}
\alias{remove_null_items}
\title{remove null items}
\usage{
remove_null_items(df_list)
}
\arguments{
\item{df_list}{A list of data frames with latitude and longitude data for each species.}
}
\value{
The same list of data frames with any elements in the list removed that were null
}
\description{
a function to remove null items
}
