% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ocean_points.R
\name{remove_ocean_points}
\alias{remove_ocean_points}
\title{remove ocean points}
\usage{
remove_ocean_points(df_list, world_map)
}
\arguments{
\item{df_list}{A list of data frames with latitude and longitude data for each species.}

\item{world_map}{A shapefile of the world's coastlines, can be obtained from the wrld_simpl data from the maptools package}
}
\value{
The same list of dataframes with localities lying over the ocean removed
}
\description{
A function to remove points in the ocean given a world map shapefile
}
