% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_perf_0_90_180.R
\name{remove_perf_0_90_180}
\alias{remove_perf_0_90_180}
\title{remove perfect 0, 90, 180 coordinates}
\usage{
remove_perf_0_90_180(df_list)
}
\arguments{
\item{df_list}{A list of data frames with latitude and longitude data for each species.}
}
\value{
The same list of data frames with any points with perfect 0, 90, or 180 coordinates removed
}
\description{
A function to remove points with perfect 0, 90, or 180 coordinates
}
