% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_points_outside_nat_range.R
\name{remove_points_outside_nat_range}
\alias{remove_points_outside_nat_range}
\title{remove points outside native range}
\usage{
remove_points_outside_nat_range(
  df_list,
  botan_map = kew_map_level_2,
  nat_range_df
)
}
\arguments{
\item{df_list}{A list of data frames with latitude and longitude data for each species.}

\item{botan_map}{A map used to calibrate the projection. Default set to kew_map_level_2}

\item{nat_range_df}{A list of shapefiles of the native ranges of the species in the df_list. Can be obtained from Kew Garden website}
}
\value{
The same list of data frames with occurrences removed that fell outside the defined native range of the species
}
\description{
A function that removes occurrence points from each data frame in the list if it lies outside a given native range
}
