% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{run}
\alias{run}
\alias{runCAD}
\alias{runExample}
\title{Run candi cad example or candi rad applications.}
\usage{
runCAD()

runExample(example)
}
\arguments{
\item{example}{chr(1) name of example file to run; default lists available examples}
}
\description{
\code{runCAD()} Launch a Shiny app to demo the CAD clinical trial app.
\code{runExample()} Lists all available example apps and runs any of them.
}
\examples{
if (interactive()) {
   runCAD()
   candi::runExample()
   candi::runExample("CANDI_CAD")
}
}
