% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio.R
\name{save_usr_input}
\alias{save_usr_input}
\title{Save a data frame with user input to csv}
\usage{
save_usr_input(x, dir)
}
\arguments{
\item{x}{data.frame to save}

\item{dir}{chr(1)}
}
\value{
called for side effect
}
\description{
The filename is computed as the md5 hash of the data.frame
}
\examples{
save_usr_input(mtcars, "~/app_data_cxrTargetDiff/usr_inpt")
}
