% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_similarImageUi.R
\name{similarImg}
\alias{similarImg}
\title{CANDI PC-based Similar Image Search Shiny Module Server}
\usage{
similarImg(input, output, session, testImgId, test_imgs_df, hist_imgs_df,
  dxs_chr = candi::candiOpt(dxs_chr),
  img_dir = candi::candiOpt(large_img_dir))
}
\arguments{
\item{input}{shiny session input}

\item{output}{shiny session output}

\item{session}{shiny session id}

\item{testImgId}{reactive expression that returns chr(1) img_id of current test image}

\item{test_imgs_df}{radiograph-major data.frame with at least img_id and PCs}

\item{hist_imgs_df}{radiograph-major data.frame with at least img_id and PCs,
and also other cols to map to scatter plot aesthetics}

\item{img_dir}{chr(1)}
}
\value{
data.frame similarImgsDf
}
\description{
CANDI PC-based Similar Image Search Shiny Module Server
}
\seealso{
\code{\link{similarImgUi}}

Other shiny_module: \code{\link{impressionInput}},
  \code{\link{impression}}, \code{\link{radiographOutput}},
  \code{\link{similarImgUi}}
}
