% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_pieces}
\alias{split_pieces}
\title{Split a string vector into a list of string piece vectors.}
\usage{
split_pieces(x_chr, split_pat = character())
}
\arguments{
\item{x_chr}{character vector of strings to be split}

\item{split_pat}{regex pattern to delimit pieces}
}
\value{
pieces list with same length as x_chr, whose i-th element contains a
character vector of split products of x_chr[i]
}
\description{
First, a replacement is performed at alphanumeric piece edges to create
pieces that are delimited by a space or punctuation character.
Second, the string is split by split_pat into pieces
}
