% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_case_title}
\alias{str_case_title}
\title{Convert string to title case.}
\usage{
str_case_title(x_chr)
}
\arguments{
\item{x_chr}{character vector}
}
\value{
x_title character vector with same length as x_chr, but title format
}
\description{
Transform a character vector into a title case representation.
First, separate camelCase and alphanumeric boundaries into spaces
Second, split pieces by symbols or spaces
Third, use \code{\link[stringr]{str_to_title}} to glue pieces into title.
}
\examples{
str_case_title(rownames(mtcars))
}
\seealso{
Other str_case: \code{\link{str_case_snake}}
}
