% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stringFunc}
\alias{stringFunc}
\title{Convert an object to a string representation with grammer for multiples}
\usage{
stringFunc(x)
}
\arguments{
\item{x}{object to compress.  if not x_chr, \code{\link[base]{format}} is used to coerce x to a string.}
}
\value{
chr(1)
}
\description{
From caret
}
\examples{
stringFunc("foo")
stringFunc(c("foo", "bar"))
lift_vd(stringFunc)("foo", "bar")
stringFunc(letters)
}
