% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t2idf}
\alias{t2idf}
\title{Transpose a df to another indexed data.frame}
\usage{
t2idf(df, name = "column", value = "value")
}
\arguments{
\item{df}{data.frame}

\item{name}{chr(1) column name for original data.frame colnames}

\item{value}{chr(1) colname for original row 1 values}
}
\value{
data.frame
}
\description{
The base \code{\link[base]{t}} produces a matrix.
This wrapper calls \code{\link[base]{t}} followed by
\code{\link{mtx2idf}} to produce a data.frame instead
}
\examples{
summarise_if(mtcars, .p=is.numeric, .f=mean) \%>\% t2idf()
summarise_if(mtcars, .p=is.numeric, .f=mean) \%>\% t2idf("variable", "mean")
summarise_if(group_by(mtcars, cyl), .p=is.numeric, .f=mean) \%>\% t2idf("variable", "mean")
}
