% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tab.Perf.R
\name{Tab.Perf}
\alias{Tab.Perf}
\title{Tabulate Performance Scores for package models}
\usage{
Tab.Perf()
}
\value{
a \code{\link[knitr]{kable}} table with various metrics for each model
}
\description{
Tabulate Performance Scores for package models
}
\examples{
Tab.Perf()
}
