% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candi-package.R
\docType{data}
\name{test_imgs}
\alias{test_imgs}
\title{chr(113) with image identifiers to be tested}
\format{An object of class \code{character} of length 113.}
\usage{
test_imgs
}
\description{
chr(113) with image identifiers to be tested
}
\keyword{datasets}
