% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_traceOutput.R
\name{traceOutput}
\alias{traceOutput}
\title{Trace verbatim outputs of reactive variables}
\usage{
traceOutput(id)
}
\description{
Trace verbatim outputs of reactive variables
}
\examples{
if (interactive()) {
library(shiny)

shinyApp(
    ui = fluidPage(
        textInput("user_name", "Radiologist Name:", value = "Marcus"),
        traceOutput("trace")
    ),
    server = function(input, output, session) {
        callModule(trace, "trace",
                   user_nameIn = reactive(input$user_name),
                   usrImpressionDf = reactive(mtcars),
                   usageLst = reactive({
                       cdata <- session$clientData
                       cnames <- names(cdata)
                       cvals <- lapply(cnames, function(name) {cdata[[name]]})
                       cvals \%>\% purrr::set_names(cnames)
                   }))
    }
)}
}
