% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Viz.Image}
\alias{Viz.Image}
\title{Rasterize an EBImage Image object figure.}
\usage{
Viz.Image(img_arr)
}
\arguments{
\item{img_arr}{an EMIage `Image` object}
}
\value{
side-effect: draws figure
}
\description{
Rasterize an EBImage Image object figure.
}
