% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.Roc.R
\name{Viz.Roc}
\alias{Viz.Roc}
\title{Generate an ROC plot of this pkgs internal model data}
\usage{
Viz.Roc()
}
\value{
returns a ggplot obj with roc curve labels
}
\description{
Generate an ROC plot of this pkgs internal model data
}
\examples{
Viz.Roc()
}
