% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candi-package.R
\docType{data}
\name{cases}
\alias{cases}
\title{case-major data.frame with patient and diagnostic variables from IU dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3786 rows and 80 columns.}
\usage{
cases
}
\description{
source: https://openi.nlm.nih.gov/faq.php
}
\keyword{datasets}
