% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ids.R
\name{imgIds2Cases}
\alias{imgIds2Cases}
\title{Convert a vector of image IDs to corresponding unique case IDs.}
\usage{
imgIds2Cases(img_ids)
}
\arguments{
\item{img_ids}{x_chr of image ids to convert to case ids}
}
\value{
x_int of case identifiers
}
\description{
Convert a vector of image IDs to corresponding unique case IDs.
}
\examples{
imgs_avail = candiOpt(large_img_dir) \%>\% list.files() \%>\% fp_stem()
imgIds2Cases(imgs_avail)
imgIds2Cases("iu_1_1")
imgIds2Cases(1)
imgIds2Cases("foobar")
}
