% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impressionUi.R
\name{impression}
\alias{impression}
\title{CANDI User Impression Interface Module Server Function}
\usage{
impression(input, output, session)
}
\arguments{
\item{input, output, session}{shiny module server-client mgmt}
}
\value{
reactive conductor usrImpressionDf with user input values
}
\description{
CANDI User Impression Interface Module Server Function
}
\seealso{
\code{\link{impressionUi}}

Other shiny_module: \code{\link{impressionUi}},
  \code{\link{similarImgUi}}, \code{\link{similarImg}}
}
