% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impressionInput.R
\name{impression}
\alias{impression}
\title{CANDI User Impression Interface Module Server Function}
\usage{
impression(input, output, session)
}
\arguments{
\item{input, output, session}{shiny module server-client mgmt}
}
\value{
a function, wrapped in S3 class reactive.  see \code{\link[shiny]{reactive}}
}
\description{
CANDI User Impression Interface Module Server Function
}
\seealso{
\code{\link{impressionInput}}

Other shiny_module: \code{\link{impressionInput}},
  \code{\link{similarImgUi}}, \code{\link{similarImg}}
}
