% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impressionInput.R
\name{impressionInput}
\alias{impressionInput}
\title{Impression User Interface Module}
\usage{
impressionInput(id, dx_chr)
}
\arguments{
\item{id}{chr(1) namespace for the module, decided by the caller at the time the module is used}

\item{dx_chr}{chr(n) diagnoses to include in the classification input}
}
\value{
\code{\link[shiny]{tagList}} of impression ui components
}
\description{
Create a collection of ui inputs for a user to indicate binary classification of
each dx in \code{dx_chr}.
}
\examples{
impressionInput(id="imp", dx="Cardiomegaly")
}
\seealso{
\code{\link{impression}}

Other shiny_module: \code{\link{impression}},
  \code{\link{similarImgUi}}, \code{\link{similarImg}}
}
