% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio.R
\name{load_radiograph}
\alias{load_radiograph}
\title{Load radiographs into EBImage Images}
\usage{
load_radiograph(img_id, img_dir)
}
\arguments{
\item{img_id}{chr(1)}

\item{img_dir}{chr(1)}
}
\value{
\code{\link[EBImage]{Image}}
}
\description{
Wrapper for \code{\link[EBImage]{readImage}}
}
\examples{
\dontrun{
  img <- load_radiograph(img_id="iu_101_2", img_dir="/www/app_data_cxrTargetDiff/large_jpgs")
  EBImage::display(img)
}
}
