% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio_usrInput.R
\name{load_usr_input}
\alias{load_usr_input}
\title{Load input app data from a user}
\usage{
load_usr_input(user, dir)
}
\arguments{
\item{user}{chr(1) user to filter user_name column by}

\item{dir}{chr(1)}
}
\value{
data.frame
}
\description{
By default, data is loaded form a directory indicated by a
}
\examples{
\dontrun{
  load_usr_input("Marcus", dir = file.path(candiOpt(app_data_dir), "rad_case", "log"))
}
}
