% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsio_usrInput.R
\name{save_usr_input}
\alias{save_usr_input}
\title{Save a data frame with user input to csv}
\usage{
save_usr_input(x, dir, subdir = "")
}
\arguments{
\item{x}{data.frame to save}

\item{dir}{chr(1)}
}
\value{
called for side effect; but returns target fp for reference
}
\description{
A column is added with save time and the file is named by md5 hash.
}
\examples{
save_fp <- save_usr_input(mtcars, dir = tempdir())  # Save to tmp
read_csv(save_fp)
save_fp <- save_usr_input(mtcars, candiOpt(app_data_dir)); print(save_fp)  # save to usr_input_dir
read_csv(save_fp); file.remove(save_fp)            # and remove
}
