% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_similarImageUi.R
\name{similarImg}
\alias{similarImg}
\title{CANDI PC-based Similar Image Search Shiny Module Server}
\usage{
similarImg(input, output, session, testImgId, img_dir, dx_chr)
}
\arguments{
\item{input, output, session}{shiny module server-client mgmt}

\item{testImgId}{reactive expression that returns chr(1) img_id of current test image}

\item{img_dir}{chr(1)}

\item{dx_chr}{chr(n) diagnoses to incldue}
}
\value{
data.frame similarImgsDf
}
\description{
CANDI PC-based Similar Image Search Shiny Module Server
}
\seealso{
\code{\link{similarImgUi}}

Other shiny_module: \code{\link{impressionInput}},
  \code{\link{impression}}, \code{\link{radiographOutput}},
  \code{\link{similarImgUi}}
}
