% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_similarImageUi.R
\name{similarImgUi}
\alias{similarImgUi}
\title{Shiny module for PC based similar historical image lookup}
\usage{
similarImgUi(id)
}
\arguments{
\item{id}{chr(1) namespace for the module, decided by the caller at the time the module is used}
}
\value{
\code{\link[shiny]{tagList}} of similar image search ui components
}
\description{
Shiny module for PC based similar historical image lookup
}
\examples{
if (interactive()) {
library(shiny)

shinyApp(
    ui = fluidPage(
        similarImgUi("similarImg")
    ),
    server = function(input, output) {
        callModule(similarImg, "similarImg",
                   testImgId = "iu_1_1",
                   img_dir = candiOpt(large_img_dir),
                   dx_chr = candiOpt(dxs_chr))
    }
)}
}
\seealso{
\code{\link{similarImg}}

Other shiny_module: \code{\link{impressionInput}},
  \code{\link{impression}}, \code{\link{similarImg}}
}
