% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_similarImageUi.R
\name{similarImgUi}
\alias{similarImgUi}
\title{Shiny module for PC based similar historical image lookup}
\usage{
similarImgUi(id)
}
\arguments{
\item{id}{chr(1) namespace for the module, decided by the caller at the time the module is used}
}
\value{
\code{\link[shiny]{tagList}} of similar image search ui components
}
\description{
Shiny module for PC based similar historical image lookup
}
\examples{
if (interactive()) {

library(shiny)

kN_HIST_IMGS <- 100
test_imgs <- candi::radiographs \%>\%
    filter(img_id \%in\% candi::test_imgs) \%>\%
    with_sep(left_join, cases, by="case")
hist_imgs <- radiographs \%>\%
    filter(img_id \%ni\% candi::test_imgs) \%>\%
    sample_n(kN_HIST_IMGS) \%>\%
    with_sep(left_join, candi::cases, by="case")

shinyApp(
    ui = fluidPage(
        useShinyjs(),
        selectInput("imgIdIn", "Image Id:", choices = test_imgs$img_id),
        similarImgUi("similarImg")
    ),
    server = function(input, output, session) {
        callModule(similarImg, "similarImg",
                   testImgId = reactive(input$imgIdIn),
                   test_imgs_df = test_imgs,
                   hist_imgs_df = hist_imgs)
    }
)
}
}
\seealso{
\code{\link{similarImg}}

Other shiny_module: \code{\link{impressionInput}},
  \code{\link{impression}}, \code{\link{radiographOutput}},
  \code{\link{similarImg}}
}
