% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_to_shapefile.R
\name{trim_to_shapefile}
\alias{trim_to_shapefile}
\title{trim to shapefile}
\usage{
trim_to_shapefile(environmental_raster, cookie_cutter)
}
\arguments{
\item{environmental_raster}{A raster or raster stack}

\item{cookie_cutter}{A geographic shapefile provided by the user to trim the raster(s) to}
}
\description{
A function to trim a stack of environmental data to a given shapefile
}
